requirements.txt
streamlit>=1.24.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
bitsandbytes>=0.39.0
pillow>=9.0.0
